Changed migration file to bigint.

+++ db/migrate/20190611095144_create_family_plannings.rb
@@ -2,8 +2,8 @@ class CreateFamilyPlannings < ActiveRecord::Migration[5.2]
   def change
     create_table   :family_plannings , :primary_key => :family_planning_id do |t|
     	t.bigint      :encounter_id
-    	t.integer      :concept_id
-    	t.integer      :value_coded
+    	t.bigint      :concept_id
+    	t.bigint      :value_coded
     	t.boolean    :voided, null: false, default: 0
     	t.bigint       :voided_by
     	t.datetime   :voided_date
