+++ db/migrate/20190611095419_create_master_definitions.rb
@@ -4,7 +4,7 @@ class CreateMasterDefinitions < ActiveRecord::Migration[5.2]
           t.string    :definition
           t.string     :description
           t.boolean   :voided, null: false, default: 0
-          t.integer    :voided_by
+          t.bigint    :voided_by
           t.datetime   :voided_date
           t.string     :void_reason
